Working plotting (tested on VGG16)
